refactor: Refactor tree context and more minor changes
Delete Schema context and create Trees context to replace it, so it will
be more aligned to the namespaces standards.

Implement password hash function to hash the user's password when user is
created or password is changed.

Implement several new tests to cover more cases.

Modify all tests to use ex_machina to make it easier to create test
data and associations.

Create a docs folder to store documentation and add API Specification in
html and yaml formats.