refactor: Enable PostGIS
We will use PostGIS to make it easier to work with geographic points

- Changes in config files to support geo_postgis

- Changed postgres image in docker-compose.yml to postgis image

- Created a new migration to enable postgis extension, to remove
longitude and latitude fields in tree schema and to create a
postgis geometry point called location

- Created two new functions inside tree schema file, one to verify if
longitude and latitude inside the postgis geometry object are numbers
and other to verify if the user_id passed on a update call is not nil

- Created a new function inside tree_controller to extract both lat
and lng fields from the request body, build a geometry point and add
it in params as location. This function is called when a create or an
update happens.

- Updated tree's view to return data as defined in the API Spec

Referece used: https://akoutmos.com/post/prometheus-postgis-and-phoenix/#step-2-set-up-basic-docker-compose-stack