doc: Update API Specification
Changes made in tree info returned by the API, changes will not
impact the front-end but will make the back-end job much easier.

We removed the fields 'owner' and 'owner_nickname' and added an object
'owner' that contains two fields, 'id' and 'nickname'. Also, when you
request information about a tree using the user id
(/users/:userId/trees/:treeId), it will not return the new 'owner'
object because it makes no sense to return information about a user that
the requester already knows.